批量大包代码

+++ case/boxhelper/a00_boxhelper.py
@@ -25,8 +25,9 @@ def kill_exe(_exe):
     tmp_print(f'{_exe}已全部关闭!')
 
 def main():
+    print('当前辅助器版本号: V', boxhelper_version)
     # 检查网络
-    is_network = check_all_nets(is_adb=False)
+    is_network = check_all_nets()
     # 如果网络正常
     if is_network:
         # 查询minio的母盒包(此处无需再检查是否有新版本,只要进入到这里,就说明有新版本)
@@ -49,21 +50,26 @@ def main():
             # 下载最新母盒包(下载到用户目录, 因为打包上传时直接压缩build文件夹)
             # .xxx/build/exe.win-arm64-3.11/motherbox_1002.zip
             local_zip = os.path.join(build_in_who, f'motherbox_{new_version}.zip')
-            download_obj(local_zip, motherbox_newpath)
-            # 解压母盒包
-            tmp_print(f'正在解压母盒包...')
-            shutil.unpack_archive(str(local_zip), build_in_who)
-            # 删除压缩包
-            tmp_print(f'正在删除压缩包...')
-            os.remove(str(local_zip))
-            # 启动新的母盒
-            tmp_print(f'正在启动新母盒...')
-            newbox_exe = subprocess.Popen([exe_abs_path], creationflags=subprocess.CREATE_NEW_CONSOLE)
-            tmp_print("当前母盒辅助器进程ID:", newbox_exe.pid)
-            # 退出母盒辅助器
-            tmp_print(f'正在退出母盒辅助器...')
-            exit(0)
-            sys.exit(0)
+            is_downed = download_obj(local_zip, motherbox_newpath)
+            if not is_downed:
+                tmp_print(f'下载最新母盒包失败, 请检查网络连接!')
+                input('请确保网络连接正常, 并按任意键再尝试')
+                main()
+            else:
+                # 解压母盒包
+                tmp_print(f'正在解压母盒包...')
+                shutil.unpack_archive(str(local_zip), build_in_who)
+                # 删除压缩包
+                tmp_print(f'正在删除压缩包...')
+                os.remove(str(local_zip))
+                # 启动新的母盒
+                tmp_print(f'正在启动新母盒...')
+                newbox_exe = subprocess.Popen([exe_abs_path], creationflags=subprocess.CREATE_NEW_CONSOLE)
+                tmp_print("当前母盒进程ID:", newbox_exe.pid)
+                # 退出母盒辅助器
+                tmp_print(f'正在退出母盒辅助器...')
+                exit(0)
+                sys.exit(0)
 
         else:
             tmp_print(f'未找到{motherbox_exe}所在目录!')
@@ -75,7 +81,6 @@ def main():
         main()
 
 if __name__ == '__main__':
-    # main()
     # shutil.rmtree(r'C:\Users\huilin.xu\Desktop\d\build', onerror=del_rw)
 
     # 下载 - 解压 - 删除
@@ -89,6 +94,5 @@ if __name__ == '__main__':
     # build_dir = os.path.dirname(str(arm64_dir))  # build
     # tmp_print(f'正在删除当前母盒...')
     # shutil.rmtree(build_dir, onerror=del_rw)
-    # main()
-    check_all_nets()
+    main()
     pass
